# docs/api/openapi.yaml
openapi: '3.0.3'
info:
  title: Kalki-Mirror API
  version: 0.2.1
  description: >-
    REST API for the KALKI Tantrik Intelligence Platform.
    Pattern recognition, transit geometry, AI-powered Vedic content,
    and membership management.
  contact:
    name: KALKI Platform
    url: https://www.astrokalki.com
  license:
    name: Proprietary
servers:
  - url: https://www.astrokalki.com
    description: Production
...
paths:
...
  /api/ai/ask:
    post:
      operationId: aiAsk
      tags: [AI]
      summary: Public grounded Q&A (Vol. 4 #14)
      description: >-
        Corpus-grounded question answering for the public /ask surface.
        Reuses the wizard's citation pool at the public 'prithvi' tier plus
        the deterministic pattern-bridge boost. CORPUS-OR-SILENCE contract:
        every grounded answer cites folio slugs validated against the
        retrieved set; weak retrieval or a grounding-contract breach returns
        grounded=false instead of an improvised answer. Identical
        (query · retrieved set) pairs are served from SynthesisCache. Every
        call fires an ai_ask event (Vol. 4 #17). No auth. Rate limited
        5/min/IP. Returns 503 if LLM is not configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  minLength: 3
                  maxLength: 500
      responses:
        '200':
          description: Grounded answer with validated citations, or explicit silence
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: Grounded answer
                    required: [grounded, answer, citations, model, method, cached]
                    properties:
                      grounded:
                        type: boolean
                        enum: [true]
                      answer:
                        type: string
                      citations:
                        type: array
                        items:
                          type: object
                          properties:
                            slug:
                              type: string
                            section:
                              type: string
                            similarity:
                              type: number
                      model:
                        type: string
                      method:
                        type: string
                        enum: [embedding, keyword]
                      cached:
                        type: boolean
                  - type: object
                    description: Corpus-or-silence refusal
                    required: [grounded, reason]
                    properties:
                      grounded:
                        type: boolean
                        enum: [false]
                      reason:
                        type: string
                        enum: [corpus_silent, ungrounded_output]
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/ai/consultation-screen:
    post:
      operationId: aiConsultationScreen
      tags: [AI]
      summary: AI consultation screening
      description: >-
        Screens a user's consultation message before routing.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, message]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 200
                message:
                  type: string
                  minLength: 10
                  maxLength: 5000
      responses:
        '200':
          description: Screening result
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/ai/breathwork:
    post:
      operationId: aiBreathwork
      tags: [AI]
      summary: AI breathwork session generator
      description: >-
        Generates a guided breathwork session of the specified type and duration.
        No auth. Rate limited 5/min/IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum:
                    - calming
                    - energizing
                    - focus
                    - nadi-shuddhi
                    - bhramari
                    - custom
                duration:
                  type: integer
                  minimum: 3
                  maximum: 60
                  default: 15
                  description: Session duration in minutes
      responses:
        '200':
          description: Breathwork session generated
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT via Bearer token (cookie-based).
        Used for standard authenticated routes.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT with ADMIN, SUPERADMIN, EDITOR, or REVIEWER role required.

  # ── Reusable Responses ──────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: LLM not configured or returned invalid JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Health
...